api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.40.0
ollama>=0.6.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
alembic>=1.13.0
celery[redis]>=5.3.0
cryptography>=42.0.0
